2019 Day 16 Part 2
Thanks to some old reddit hint threads. In the second half of the input the pattern is just 0...0,1...1. So we are just summing the back half or less of the digits and then doing a mod 10.

In addition, the offset added means we only need to calculate the digits after the offset, anything before it gets zeroed out by the pattern.