2021 Day 24 Part 1 + 2
Optimized as much as I care to do. Part 1 runs super slow because the given input answer starts with a 3. Part 2 actually only takes about 6 seconds. I think if I parallelized it to run all digits (or like the first 5) at once and exit when a solution is found it might speed up more.