2022 Day 11
Meh. Don't really care for the math type puzzles. Was on the right track of modding the value before sending it but had mostly tested doing things with only the next monkey in line. It's late, I'm hungry, just looked up the answer. Then wasted time not resetting the classes between parts.