2022 Day 13
Spent way too long debugging a possible parsing bug for it to have been somebody on reddit mis typing a sample test case. Turns out you can parse like JSON real easy.

Had a bug in the comparison code too, not sure what was wrong with it. Seemed to line up exactly with the code online but when I switched to numerical instead of bool? things started working. Fun instance of all the samples working but the real input not on a problem where it's hard to find which input is the problem.

Part 2 was super easy, create an IComparer and throw in the existing validator. Then fix the off by one error because who uses 1 indexed items.