2022 Day 17 Part 2
The 1 trillion thing makes it pretty obvious you aren't meant to simulate the whole thing. Finding a cycle isn't too hard once you come up with a good hash for it (or at least one that works on the example and your input). The math on the cycle height was a little annoying with adding 1 to get height vs Y coord. Doesn't run super fast, about 4 seconds and change. Could probably improve with a better hashing func.