2022 Day 19 Part 1 and 2
Another puzzle I'm not really a fan of. The ones where it is just some kind of search and then you have to keep pruning to be effectient just don't do much for me. Especially given examples like day 16 where the best optimizations have nothing to do with logic and everything to do with how you store info. Basically turns in to make it was less readable so it can be faster.