2022 Day 24 Part 2
About 1 sec for part 1 and 1.7 for part 2. Really really surprised that it worked that easily. The only bug from copying the part 1 code was not allowing a move back to the start or allowing us to go south from the end.

Other than that it worked perfectly, not sure if it was actually this easy or my part one was already using whatever trick part 2 was supposed to need.