2023 Day 5 Part 2
Hard part 2 for day 5. First time I really felt like I was stuck for a while this year. First thought was just iterating the location and seeing if it mapped back to a valid seed range. But the answer ended up being in the 500M range and that was too slow. Looked on reddit for approaches other people used and saw somebody mention they converted the input ranges to the next stage. That is what this solution is, intuitively not that complex but it was a pain to write.