2023 Day 6 Part 2
A much easier day than yesterday or I just got lucky. Started part 1 with something other than "check every value" because that never works for part 2 and wasn't that hard to write. The bigger surprise was part 2 working with that code and relatively fast (<200 ms). I could probably speed it up by making it more of a search on both bounds but this is good nuff.