2023 Day 8 Part 2
Not as hard as initially thought. Did a quick check on brute force (too slow) so then I wrote the foreach loop for part 2. My big mistake there was stopping the loop & saving the loop size as soon as we re-entered a node instead of waiting to return to the goal. So the resulting numbers in loopData made no sense at first. Once I fixed that and saw every start point looped to the end with one value it was just grab an LCM formula and plug em in.