2023 Day 10 Part 2
Slow currently. Also had to look up hints on part 2. This solution makes the most sense to me. We're counting vertical walls of the loop passed so far. If the number is odd we are outside the loop. Added tricky part that a few of the corner pieces combined is virtually a vertical wall.