2023 Day 12 Part 2
Worked after I upped the counting var to a decimal. Instead of jumping to the next question mark we go char by char (might not have been necessary). Kept adding to the tracking state until it output the correct results for the examples. Runs in about 12 sec for part 2. There is probably a smarter DP way of doing this that does a lot less work per iteration.