2023 Day 17 Part 1
Pathfinding with extra rules was pretty hard to code up. Not sure I like the route I took of expanding the neighbors vs doing some kind of other tracking on which directions are valid. You can see from the commented code that it made it hard to debug.

First bug was not using coordinate + direction in the cost & path dictionaries. Then had a bug where we didn't add to the running cost when skipping a neighbor because it was visited which gave lower cost paths than were actually there.