2023 Day 17 Part 2
I changed my mind, I'm a genius. Part 2 is an exact copy of part 1 with so slight tweaks to how we pick neighbors. The way I coded part 1 ended up making this super easy, the only trick was making sure to still incur the cost of invalid stopping points.

Not particuallarly fast code today, 4.6s for part 1 and 29s for part 2. Not sure if that is the algorithms fault (it's more or less djikstra), how I coded it (don't think I have any list lookups or anything), or just not a fast to solve problem.