2023 Day 20 Part 2
Similar to day 8. There are special things about the input that make this solution work:
- rx is only connected to 4 inverters
- Those inverters are in separate loops
- Each of those inverters has one input

Putting that together: We can watch each inverter that goes in to rx and look for a low pulse to be sent to get the cycle time of that subsection. Once we find all four, the LCM of those cycles is what iteration we would get all four inverters sent a low signal, which becomes a high signal in the input to rx, which sends a single low to rx.

Ran everything together to not have to reset the state.