2023 Day 21 Part 2
Well I got an infinite map class out of it, and made the code faster. Funny how I hate this puzzle right after making fun of people for hating yesterdays puzzle. But in this case you basically output numbers until it looks like a quadratic I guess? And then use something else to calculate the quadratic because this isn't Python.