2024 Day 6 Part 2
So we learned json serialization is awful. Found a new package to do a binary serialization that works without errors about stuff like having coordinate in a dictionary.

Cleaning up the structure a little (a lot of puzzles need part 1 to run before part 2 so always resetting isn't ideal)

For the actual solution, it feels a little brute force but I'm not sure if you could memoize anything since the target space is always changing. Ran into an issue early continuing the path from the new block, not considering that the new block might alter the previous path