2024 Day 17 Part 2
Oh now it's getting tough. I figured out the internals of what the program was doing, but got too hung up on thinking it was some kind of reversible process. It's more like a hash so you can't just directly reverse it or if you can, not easily. After some hints from reddit came to this solution. Build up the output string from the last digit (since the later digits will not alter it) until we have found the values of A that result in the desired output. There aren't that many and the list of valid values at each step is usually under 20. Then just pick the lowest one