2024 Day 18 Part 1
Lots of edits from adding type to the load input. But a few puzzles this year have had examples where the map size is smaller than the actual, but that isn't known from just the input.

As far as the puzzle, just a simpler copy of the path finding from day 16.