2024 Day 20 Part 1
Interesting problem. First we need to find a path in reverse to get a dictionary of distance to the end. Then to simulate the cheat we can just find all valid points along that path where we could cheat. Then finding the path to end is just a simple lookup.