2024 Day 21 Part 2
Takes some inspection of the example to see the pattern. If robot 0 is controlling the numeric keypad and 1 is controlling the first directional. At robot X+1 doing repeated sequence on robot X runs the same moves over and over. So you can recurse and memoize to prevent the huge amount of possibilities showing up.