2025 Day 3 Part 2
It really wasn't that hard but it took me forever. I started with Dfs but longer and that didn't work. Memoizing it didn't help either. Then I tried working backwards in the string replacing digits but that has too many edge cases. Finally came up with this solution which is the simplest answer when you think about it. Just always picking the highest available option for the next digit, because that'll always be higher than anything else we can do