2025 Day 8 Part 2
Not too bad to code really, but the optimizations I made for part 1 definitely made it easier. Takes 16 seconds on part 2 which is pretty slow, but I don't have any obvious ideas on how to speed it up. I can't imagine that check for the circuit is the slow part