2025 Day 9 Part 2
Kind of hard? The code I ended up with was my initial thought but I didn't start on it immediately knowing that the input size was going to be too large to make it feasible. But then I read a Reddit comment about compressing the grid space (very cool) and then I could implement it. Runs in about 8 seconds so not crazy fast but not bad. I would guess the validity check on the rectangles could be improved