2025 Day 11 Part 2
Wow that was easy and I made it so hard for myself. Maybe it was because my part 1 for some reason built the entire path instead of the count and got me going down the wrong way. I spent way too much time thinking I needed to simplify the graph. Or trying to memoize but build the entire path. The only truly tricky part was the requirement in the path, but you can just use a bool to flag when we've found it and have that effect the termination.